correction to delete room

+++ ftp/views.py
@@ -492,10 +492,12 @@ class DeleteRoom(APIView):
         
         guest_session = self.request.session.session_key
 
-        host = Guest.objects.filter(guest = guest_session)[0]
-        room = Room.objects.filter(code = str(host.room))[0]
+        room = Room.objects.filter(host = guest_session)[0]
 
-        delete_room_files(room)
-        room.delete()
+        if room.exists():
+            delete_room_files(room)
+            room.delete()
+            return Response({"Room Destroyed": "All the data has been purged"}, status = status.HTTP_301_MOVED_PERMANENTLY)
 
-        return Response({"Room Destroyed": "All the data has been purged"}, status = status.HTTP_301_MOVED_PERMANENTLY)
+        else:
+            return Response({"Bad Request": "Invalid Data..."}, status = status.HTTP_400_BAD_REQUEST)
